Update custom-resource-page to use latest extensions api version
