+++ custom-resource-page/package.json
@@ -31,7 +31,7 @@
   "devDependencies": {
     "ts-loader": "^8.0.4",
     "typescript": "^4.0.3",
-    "@k8slens/extensions": "^4.0.0-alpha.5",
+    "@k8slens/extensions": "^4.0.0-rc.2",
     "@types/node": "^12.18.3",
     "@types/react": "^16.9.35",
     "mobx": "^5.15.5",
